Allow users to ssh into Cloud Foundry conainers
Added a shell script that configures the container for a unprivileges sshd. The
script downloads OpenSSH source code, patches sshd, and compiles it before
launching it to listen on port 2022.

Patching sshd is ncessary, since the daemon tries to assign `vcap:tty` to the
new PTY it creates. That is prohibited in the cflinuxfs2 container types in
Cloud Foundry since the `vcap` user is not a member of `tty` unix group.

Added manifest.yml, as needed for deploying this app in Cloud Foundry.

Added instructions to README.

Added Go vendor support so that Go buildpacks in Cloud Foundry can build this
package.