Use Ubuntu's OpenSSH server, instead of building our own
We used to build our own, because we had to patch the sshd code to
work around the fact that OpenSSH set the gid, and when it's not
able to, it exits with an error.

Now that the official OpenSSH's sshd does not exit with error, we do
not need to build our own server, and can use the official one.

Instead of downloading the .deb file from the internet, we store it in
repo here, since we have seen many times that the links to the internet
break, causing massive pain.

In passing, we also changed all relative paths to absolute paths, to
make it obvious where various files are located.

I took this opportunity to also create a Dockerfile that helps emulate
the Cloud Foundry environment, where this repo is actually used. This
Dockerfile would help us troubleshoot issues with our deploy.sh script
locally, without having to go through the time consuming process of
pushing this repo to CF and avoid the painful process of troubleshooting
problems in that environment.